python/requirements.txt
flask==3.0.3
opencv-python==4.10.0.84
numpy==2.1.1
requests==2.32.3
python-dotenv==1.0.1
Werkzeug==3.0.3